==> R_scripts/1_species_independent_clustering_Human.rmd <==
---
title: 'Species independent clustering: Human'
author: "Cy"
date: "8/27/2020"
output:
  html_document:
    df_print: paged
---


```{r load-libraries, echo=FALSE, quietly=T}
library(dplyr)
library(Seurat)
library(SeuratWrappers)
library(ggplot2)
library(cowplot)
library(scrattch.hicat)
library(feather)
library(Matrix)

source("../R_scripts/doublet.finder.r")
source("../R_scripts/hicat_modules.r")

load("../int_data/collected_data_20200818.rda")
set.seed(0)

```


```{r select-gene-list }
spec="human"


if (spec=="macaque") {
  orthos=read.csv("../data/ortholog_table_20191122.csv",as.is=T)
  nonlocgenes=orthos$rhesus_symbol[intersect(grep("^LOC",orthos$human_symbol,invert=T),grep("^LOC",orthos$rhesus_symbol))]
  keepgen=setdiff(rownames(datlist[[spec]]),setdiff(grep("^LOC|^MT-|^RPL|^RPS|^SNAR|^CT45|^MIR|^MTND|^TRNA",rownames(datlist[[spec]]),val=T),nonlocgenes))
  datlist[[spec]]=datlist[[spec]][keepgen,]
  datlist[[spec]]@x=datlist[[spec]]@x/rep.int(Matrix::colSums(datlist[[spec]]),diff(datlist[[spec]]@p))*10^6
} else {
  if (spec=="human") {
    keepgen=grep("^LOC|^MT-|^RPL|^RPS|^SNAR|^CT45|^MIR|^MTND|[0-9]P$|-AS|-PS|DUX4L",rownames(datlist[[spec]]),val=T,invert=T)
    keepdons=which(metalist$human$external_donor_name_label %in% c("H200.1023","H200.1025","H200.1030"))
    datlist[[spec]]=datlist[[spec]][keepgen,keepdons]
    datlist[[spec]]@x=datlist[[spec]]@x/rep.int(Matrix::colSums(datlist[[spec]]),diff(datlist[[spec]]@p))*10^6
  } else {
    keepgen=grep("^LOC|^mt-|^Rpl|^Rps",rownames(datlist[[spec]]),val=T,invert=T)
  }
}
```


### calculate doublet scores ###
```{r doublet-score}
alldoublet <- doubletFinder(datlist[[spec]], select.genes=keepgen, 
                            proportion.artificial = 0.2, plot=T)
doublet.scores <- alldoublet[[1]] 


### inspect doublet.score threshold ###
alldoublet[[3]]

noise <- intersect(names(doublet.scores)[doublet.scores>0.25],colnames(datlist[[spec]]))

print(paste("Failed doublet.score threshold (0.25) =",length(noise)))


```

```{r, eval=FALSE}
save(doublet.scores,file=paste0("../analysis_output/doublet.score_", spec,".rda"))

ggsave(alldoublet[[3]], file=paste0("../analysis_output/doublet.score_",spec,".pdf"), height = 4, width = 5)
```


#clustering of cells from all 3 Human donors

_____________________________________________________


This is an initial run to determine if there are any batch/donor effects to take into account

```{r}
load(paste0("../analysis_output/doublet.score_", spec,".rda"))

keepcells1=intersect(names(doublet.scores)[doublet.scores<0.25],colnames(datlist[[spec]]))

set.seed(0)

m.int <- CreateSeuratObject(counts = datlist[[spec]][keepgen,keepcells1], meta.data = metalist[[spec]][keepcells1,])
m.int <- NormalizeData(object = m.int)

m.int <- ScaleData(object = m.int, verbose = FALSE)

m.int <- FindVariableFeatures(object = m.int, selection.method = "vst", 
                              nfeatures = 3000) #2000 for mouse and macaque
m.int <- RunPCA(object = m.int, npcs = 30, verbose = FALSE)
m.int <- RunTSNE(object = m.int, reduction = "pca", dims = 1:30)
m.int <- RunUMAP(object = m.int, reduction = "pca", dims = 1:30)
m.int <- FindNeighbors(object = m.int, k.param = 8,dims=1:30)
m.int <- FindClusters(object = m.int, resolution = 0.8) ##0.2 for macaque,0.4 for mouse
```

```{r m.int plotting all}

DimPlot(m.int,reduction = "umap",label = TRUE)
DimPlot(m.int,reduction = "umap",group.by="external_donor_name_label",label = TRUE)
DimPlot(m.int,reduction = "umap",split.by="external_donor_name_label",label = TRUE)


FeaturePlot(m.int,features=c("GAD2","SLC17A6","EBF1", "FOXP2"))
plot<- DotPlot(m.int,features=c("GAD2","SLC17A6","SNAP25","AQP4","AIF1","OLIG2","OPALIN","LAMP5", "PRKCG", "EBF1", "FOXP2", "BTNL9"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

plot<-DotPlot(m.int,features=c("GAD2","SLC17A6","SNAP25","AQP4","AIF1","OLIG2","OPALIN","LAMP5","TMEM119","MS4A7"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

table(m.int@active.ident)
```

Neuronal clusters == 0, 1, 3, 4, 6

Non-neuronal clusters == 2, 5, 7, 8

## No-Glia clustering 
```{r clustering no-glia}
keepcells2=names(m.int@active.ident)[m.int@active.ident %in% c(0, 1, 3, 4, 6)]

m.int2 <- CreateSeuratObject(counts = datlist[[spec]][keepgen,keepcells2], meta.data = metalist[[spec]][keepcells2,])
m.int2 <- NormalizeData(object = m.int2)
m.int2 <- ScaleData(object = m.int2, verbose = FALSE)
m.int2 <- FindVariableFeatures(object = m.int2, selection.method = "vst", 
                              nfeatures = 3000) #2000 for mouse and macaque
m.int2 <- RunPCA(object = m.int2, npcs = 30, verbose = FALSE)
set.seed(0)
m.int2 <- RunTSNE(object = m.int2, reduction = "pca", dims = 1:30)
m.int2 <- RunUMAP(object = m.int2, reduction = "pca", dims = 1:30)
m.int2 <- FindNeighbors(object = m.int2, k.param = 8,dims=1:30)
m.int2 <- FindClusters(object = m.int2, resolution = 0.4)

```


```{r plotting no glia }
#DimPlot(m.int, group.by = c("external_donor_name", "ident", "roi"), ncol = 3)
DimPlot(m.int2, group.by = "external_donor_name_label")
DimPlot(m.int2, group.by = "ident")
DimPlot(m.int2, group.by = "roi_label")

FeaturePlot(m.int2,features=c("GAD2","SLC17A6","OLIG2", "MBP"))
plot<- DotPlot(m.int2,features=c("GAD2","SLC17A6","SNAP25","AQP4","AIF1","OLIG2","OPALIN","MBP","LAMP5", "PRKCG", "EBF1", "FOXP2", "BNTL9"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))


m.int2$Genes.Detected_label <- as.numeric(as.character(m.int2$Genes.Detected_label))
m.int2$total_reads_label <- as.numeric(as.character(m.int2$total_reads_label))

VlnPlot(object = m.int2, features = c("Genes.Detected_label", "total_reads_label"))

table(m.int2@active.ident)
```

glutamatergic clusters == c(0,1)

gabaergic clusters == c(2,3)


## glutamatergic neurons
```{r clustering Glut}
keepcells3=names(m.int2@active.ident)[m.int2@active.ident %in% c(0,1)]


m.int3 <- CreateSeuratObject(counts = datlist[[spec]][keepgen,keepcells3], meta.data = metalist[[spec]][keepcells3,])
m.int3 <- NormalizeData(object = m.int3)
m.int3 <- ScaleData(object=m.int3,verbose=FALSE)#,vars.to.regress = "external_donor_name_label")

dimval=30

m.int3 <- FindVariableFeatures(object = m.int3, selection.method = "vst", 
                               nfeatures = 3000) #2000 for macaque and mouse
m.int3 <- RunPCA(object = m.int3, npcs = 30, verbose = FALSE)
set.seed(0)
m.int3 <- RunTSNE(object = m.int3, reduction = "pca", dims = 1:dimval)
m.int3 <- RunUMAP(object = m.int3, reduction = "pca", dims = 1:dimval)
m.int3 <- FindNeighbors(object = m.int3, k.param = 8,dim=1:dimval)
m.int3 <- FindClusters(object = m.int3, resolution = 0.4,dims=1:dimval) #0.2 

```


```{r plotting glut }
#DimPlot(m.int, group.by = c("external_donor_name", "ident", "roi"), ncol = 3)
DimPlot(m.int3, group.by = "external_donor_name_label")
DimPlot(m.int3, group.by = "ident")
DimPlot(m.int3, group.by = "roi_label")

FeaturePlot(m.int3,features=c("GAD2","SLC17A6","BTNL9", "CASQ2"))


plot<- DotPlot(m.int3,features=c("GAD2","SLC17A6","SNAP25","PRKCG","SGCD","CASQ2","SYNDIG1L","EBF1","FOXP2","LHX9","CD5","BTNL9", "BCHE","GRB14"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

m.int3$Genes.Detected_label <- as.numeric(as.character(m.int3$Genes.Detected_label))
m.int3$total_reads_label <- as.numeric(as.character(m.int3$total_reads_label))

VlnPlot(object = m.int3, features = c("Genes.Detected_label", "total_reads_label"))


table(paste0(m.int3@meta.data$external_donor_name_label,m.int3@meta.data$roi_label),m.int3@active.ident)

table(m.int3@meta.data$external_donor_name_label,m.int3@active.ident)

```

MP == c(0)

Konio/LP == c(1)


## MP
```{r mp clustering}
keepcells4=names(m.int2@active.ident)[m.int2@active.ident %in% c(0)]

m.int4 <- CreateSeuratObject(counts = datlist[[spec]][keepgen,keepcells4], meta.data = metalist[[spec]][keepcells4,])
m.int4 <- NormalizeData(object = m.int4)
m.int4 <- ScaleData(object = m.int4, verbose=FALSE,vars.to.regress = "external_donor_name_label")

m.int4 <- FindVariableFeatures(object = m.int4, selection.method = "vst", 
                               nfeatures = 3000)
m.int4 <- RunPCA(object = m.int4, npcs = 30, verbose = FALSE)
set.seed(0)
m.int4 <- RunTSNE(object = m.int4, reduction = "pca", dims = 1:30)
m.int4 <- RunUMAP(object = m.int4, reduction = "pca", dims = 1:30)
m.int4 <- FindNeighbors(object = m.int4, k.param = 8, dims = 1:30)
m.int4 <- FindClusters(object = m.int4, resolution = 0.4) 
```


```{r plotting mp }
#DimPlot(m.int, group.by = c("external_donor_name", "ident", "roi"), ncol = 3)
DimPlot(m.int4, group.by = "external_donor_name_label")
DimPlot(m.int4, group.by = "ident")
DimPlot(m.int4, group.by = "roi_label")

FeaturePlot(m.int4,features=c("GAD2","SLC17A6","BTNL9", "CASQ2"))

plot <- DotPlot(m.int4,features=c("GAD2","SLC17A6","SNAP25","FOXP2","CRH","SYT2","BTNL9","PRKCH"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

m.int4$Genes.Detected_label <- as.numeric(as.character(m.int4$Genes.Detected_label))
m.int4$total_reads_label <- as.numeric(as.character(m.int4$total_reads_label))

VlnPlot(object = m.int4, features = c("Genes.Detected_label", "total_reads_label"))

```


## Konio/LP
```{r konio/LP clustering}

keepcells5=names(m.int2@active.ident)[m.int2@active.ident %in% c(1)] 
m.int5 <- CreateSeuratObject(counts = datlist[[spec]][keepgen,keepcells5], meta.data = metalist[[spec]][keepcells5,])
m.int5 <- NormalizeData(object =m.int5)
m.int5 <- ScaleData(object =m.int5, verbose = FALSE)

m.int5 <- FindVariableFeatures(object =m.int5, selection.method = "vst", 
                                top.genes = 2000) #2000 for macaque
m.int5 <- RunPCA(object =m.int5, npcs = 30, verbose = FALSE)
set.seed(0)
m.int5 <- RunTSNE(object =m.int5, reduction = "pca", dims = 1:30)
m.int5 <- RunUMAP(object =m.int5, reduction = "pca", dims = 1:15) #30 for macaque and mouse, 5 for human
m.int5 <- FindNeighbors(object =m.int5, k.param = 8, dims = 1:30)
m.int5 <- FindClusters(object =m.int5, resolution = 0.9) 


```


```{r plotting konio/LP }
#DimPlot(m.int, group.by = c("external_donor_name", "ident", "roi"), ncol = 3)
DimPlot(m.int5, group.by = "external_donor_name_label")
DimPlot(m.int5, group.by = "ident")
DimPlot(m.int5, group.by = "roi_label")

FeaturePlot(m.int5,features=c("PENK","BCHE","CALB1", "GRB14"))
plot <- DotPlot(m.int5,features=c("GAD2","SLC17A6","SNAP25","PRKCG","SGCD","CASQ2","IL15RA","KCNH3","SYNDIG1L","EBF1","FOXP2","AQP4","KCNJ2","HAPLN2","CWH43","MLPH"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

plot <- DotPlot(m.int5,features=c("GAD2","SLC17A6","SNAP25","PRKCG","SGCD","BCHE","GRB14"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

FeaturePlot(m.int5,features=c("PRKCG","BCHE","GRB14","PENK"))

m.int5$Genes.Detected_label <- as.numeric(as.character(m.int5$Genes.Detected_label))
m.int5$total_reads_label <- as.numeric(as.character(m.int5$total_reads_label))

VlnPlot(object = m.int5, features = c("Genes.Detected_label", "total_reads_label"))

table(m.int5@active.ident)

```


##GABAergic clustering 
```{r clustering gaba}
keepcells=names(m.int2@active.ident)[m.int2@active.ident %in% c(2,3)]

m.int6 <- CreateSeuratObject(counts = datlist[[spec]][keepgen,keepcells], meta.data = metalist[[spec]][keepcells,])
m.int6 <- NormalizeData(object = m.int6)
m.int6 <- ScaleData(object = m.int6, verbose = FALSE)
m.int6 <- FindVariableFeatures(object = m.int6, selection.method = "vst", nfeatures = 3000) #2000 for mouse and macaque
m.int6 <- RunPCA(object = m.int6, npcs = 30, verbose = FALSE)
set.seed(0)
m.int6 <- RunTSNE(object = m.int6, reduction = "pca", dims = 1:10)
m.int6 <- RunUMAP(object = m.int6, reduction = "pca", dims = 1:10)
m.int6 <- FindNeighbors(object = m.int6, k.param = 5,dims=1:8)
m.int6 <- FindClusters(object = m.int6, resolution = 0.6)

table(m.int6@active.ident)
```


```{r plotting gaba }
#DimPlot(m.int, group.by = c("external_donor_name", "ident", "roi"), ncol = 3)
DimPlot(m.int6, group.by = "external_donor_name_label")
DimPlot(m.int6, group.by = "ident")
DimPlot(m.int6, group.by = "roi_label")

FeaturePlot(m.int6,features=c("PVALB","NXPH2","LAMP5","HGF"))
plot<- DotPlot(m.int6,features=c("GAD2","LAMP5","RELN","TRPC4","ALCAM","NTRK1","COL17A1","CTXN3", "HGF","PVALB","CALB1"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

m.int6$Genes.Detected_label <- as.numeric(as.character(m.int6$Genes.Detected_label))
m.int6$total_reads_label <- as.numeric(as.character(m.int6$total_reads_label))

VlnPlot(object = m.int6, features = c("Genes.Detected_label", "total_reads_label"))

print(table(m.int6@meta.data$cluster_label,m.int6@active.ident))
table(m.int6@active.ident)
```


## Glia clustering 
```{r clustering glia}
keepcells=names(m.int@active.ident)[m.int@active.ident %in% c(2, 5, 7, 8)]

m.int7 <- CreateSeuratObject(counts = datlist[[spec]][keepgen,keepcells], meta.data = metalist[[spec]][keepcells,])
m.int7 <- NormalizeData(object = m.int7)
m.int7 <- ScaleData(object = m.int7, verbose = FALSE)
m.int7 <- FindVariableFeatures(object = m.int7, selection.method = "vst", nfeatures = 3000) #2000 for mouse and macaque
m.int7 <- RunPCA(object = m.int7, npcs = 30, verbose = FALSE)
set.seed(0)
m.int7 <- RunTSNE(object = m.int7, reduction = "pca", dims = 1:30)
m.int7 <- RunUMAP(object = m.int7, reduction = "pca", dims = 1:30)
m.int7 <- FindNeighbors(object = m.int7, k.param = 8,dims=1:30)
m.int7 <- FindClusters(object = m.int7, resolution = 0.6)

```


```{r plotting glia }
#DimPlot(m.int, group.by = c("external_donor_name", "ident", "roi"), ncol = 3)
DimPlot(m.int7, group.by = "external_donor_name_label")
DimPlot(m.int7, group.by = "ident")
DimPlot(m.int7, group.by = "roi_label")

FeaturePlot(m.int7,features=c("AQP4","GFAP","AIF1","TMEM119", "MBP", "OLIG2"))


plot<- DotPlot(m.int7,features=c("AQP4","GFAP","AIF1","TMEM119","OLIG1","OLIG2","OPALIN","MOG","MBP","PDGFRA","VCAN","CSPG4"))
plot + theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

m.int7$Genes.Detected_label <- as.numeric(as.character(m.int7$Genes.Detected_label))
m.int7$total_reads_label <- as.numeric(as.character(m.int7$total_reads_label))

VlnPlot(object = m.int7, features = c("Genes.Detected_label", "total_reads_label"))

```

```{r, fig.height=8}

FeaturePlot(m.int7,features=c("OLIG1","OLIG2","OPALIN","MBP","PDGFRA","VCAN","CSPG4", "GFAP", "AQP4"))

```



```{r, echo=FALSE, eval = FALSE}
save(m.int,m.int2,m.int3,m.int4, m.int5, m.int6,m.int7, file="../int_data/Clustering_Human.rda")
```


m.int4 - cl0 == MP

m.int5 - cl0 == K1

m.int5 - cl1 == K2

m.int6 - cl0 == GABA1

m.int6 - cl1 == GABA2

m.int6 - cl2 == GABA3

m.int7 - cl0 == Oligo

m.int7 - cl1 == Astro

m.int7 - cl2 == OPC

m.int7 - cl3 == Micro
